tests/wpt/web-platform-tests/webrtc/historical.html

<!doctype html>
<title>Historical WebRTC features</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_false("reliable" in RTCDataChannel.prototype);
}, "RTCDataChannel member reliable should not exist");

[
  "addStream",
  "createDTMFSender",
  "getLocalStreams",
  "getRemoteStreams",
  "getStreamById",
  "onaddstream",
  "onremovestream",
  "removeStream",
  "updateIce",
].forEach(function(name) {
  test(function() {
    assert_false(name in RTCPeerConnection.prototype);
  }, "RTCPeerConnection member " + name + " should not exist");
});

[
  "setDirection",
].forEach(function(name) {
  test(function() {
    assert_false(name in RTCRtpTransceiver.prototype);
  }, "RTCRtpTransceiver member " + name + " should not exist");
});

[
  "DataChannel",
  "mozRTCIceCandidate",
  "mozRTCPeerConnection",
  "mozRTCSessionDescription",
  "webkitRTCPeerConnection",
].forEach(function(name) {
  test(function() {
    assert_false(name in window);
  }, name + " interface should not exist");
});
</script>
